Lec 2
JavaScript code is executed in a different way.It creates a global execution context and allocates memory to all the variables and functions even before a single line of code is executed.
we have already reserved the memory for the a for e.g var a=7;
if we hold before 7 and check it with the help of debugger so it will helps us to understand the term 'undefined' that means the memory is allocated for a although it is not defined yet. 

undefined is not equal to empty. undefined is a special keyword. it takes up its own memory but you can assume it to be like a placeholder which is kept for a time being until the variable is assigned some other value.


JavaScript is a loosely typed language.